find window

+++ src/sample/window_process/MainWindowPane.java
@@ -3,15 +3,17 @@ package sample.window_process;
 import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.Tab;
 import javafx.scene.control.TabPane;
 import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.HBox;
 import org.apache.thrift.TException;
-import sample.AlertBox;
+import sample.Errors;
 import sample.MyTab;
 import sample.thrift.PatternModel;
 import sample.thrift.WebPatternDB;
-import sample.window_process.Window;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
@@ -23,7 +25,7 @@ import java.util.List;
 /**
  * Created by ZloiY on 3/9/2017.
  */
-public class MainWindowPane {
+public class MainWindowPane implements Errors {
     private List<PatternModel> patternList;
     private TabPane tabPane;
     private WebPatternDB.Client client;
@@ -33,13 +35,23 @@ public class MainWindowPane {
         mainPane.setCenter(tabPane);
         this.client = client;
         addAllTabs();
+        HBox functionBtnBox = new HBox();
         Button addPatternBtn = new Button("+");
-        mainPane.setRight(addPatternBtn);
+        Button searchButton = new Button("Search");
+        functionBtnBox.getChildren().addAll(addPatternBtn,searchButton);
+        mainPane.setRight(functionBtnBox);
         addPatternBtn.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent event) {
                 addPatternTab();
             }
+        });
+        searchButton.setOnAction(e->{
+            SearchWindow searchTab = new SearchWindow(this.client);
+            Tab search = new Tab("Search");
+            search.setContent(searchTab.getBorderPane());
+            tabPane.getTabs().add(search);
+            tabPane.getSelectionModel().select(search);
         });
     }
 
@@ -55,12 +67,12 @@ public class MainWindowPane {
             public void handle(ActionEvent event) {
                 PatternModel newPattern = new PatternModel();
                 if (addWindow.getNewPatternName().getText().length() > 10 || addWindow.getNewPatternName().getText().isEmpty()) {
-                    AlertBox.makeAlert("Your pattern name more than 10 characters or it's empty");
+                    new Alert(Alert.AlertType.ERROR,"Your pattern name more than 10 characters or it's empty").show();
                     return;
                 }else
                     newPattern.setName(addWindow.getNewPatternName().getText());
                 if (addWindow.getNewPatternDescription().getText().length() > 500 || addWindow.getNewPatternDescription().getText().isEmpty()){
-                    AlertBox.makeAlert("Your pattern description more than 500 characters or it's empty");
+                    new Alert(Alert.AlertType.ERROR,"Your pattern description more than 500 characters or it's empty").show();
                     return;
                 }else
                     newPattern.setDescription(addWindow.getNewPatternDescription().getText());
@@ -76,11 +88,12 @@ public class MainWindowPane {
                     }
                 }
                 try {
+                    if (client.isConnected())
                     client.addPattern(newPattern);
                     tabPane.getTabs().remove(tab);
                     addAllTabs();
                 }catch (TException e){
-                    e.printStackTrace();
+                    new Alert(Alert.AlertType.ERROR,"Service is offline try again latter.("+addErr+")").show();
                 }
             }
         });
@@ -89,11 +102,13 @@ public class MainWindowPane {
     private ArrayList<PatternModel> searchAllPatterns(){
         PatternModel pattern = new PatternModel();
         try {
+            if (client.isConnected())
             return new ArrayList<>(client.findPattern(pattern));
         }catch (TException e){
-            e.printStackTrace();
+            new Alert(Alert.AlertType.ERROR,"Service is offline try again later.("+searchAllErr+")").show();
             return null;
         }
+        return null;
     }
 
     private void addAllTabs(){
@@ -117,10 +132,11 @@ public class MainWindowPane {
                     PatternModel deletePattern = new PatternModel();
                     deletePattern.setId(window.getPatternID());
                     try {
+                        if (client.isConnected())
                         client.deletePattern(deletePattern);
                         tabPane.getTabs().remove(tab);
                     }catch (TException e){
-                        e.printStackTrace();
+                        new Alert(Alert.AlertType.ERROR, "Service is offline try again later.("+deleteErr+")").show();
                     }
                 }
             });
@@ -147,14 +163,16 @@ public class MainWindowPane {
                             ImageIO.write(bufferedImage, "png", output);
                             newPattern.setSchema(output.toByteArray());
                             output.close();
+                            if (client.isConnected())
                             client.replacePattern(oldPattern,newPattern);
+                            if (client.isConnected())
                             oldPattern = client.getLastPattern();
                             Window newWindow = new Window(oldPattern);
                             newWindow.showLayout();
                             tab.setContent(newWindow.getBorderPane());
                             tab.setText(newWindow.getPatternName().getText());
                         }catch (IOException|TException exception){
-                            exception.printStackTrace();
+                            exception.getCause();
                         }
                     });
                     tab.setContent(editWindow.getBorderPane());
